Did a lot of work with the wallet

+++ main/java/com/minelife/economy/BillHandler.java
@@ -16,6 +16,6 @@ public abstract class BillHandler {
 
     public abstract void update();
 
-    public abstract void pay(Billing.Bill bill, double amount);
+    public abstract void pay(Billing.Bill bill, int amount);
 
 }
